+++ tests/notebooks/mirror/ipynb_to_percent/convert_to_py_then_test_with_update83.py
@@ -10,7 +10,7 @@
 # ---
 
 # %%
-%%time
+# %%time
 
 print('asdf')