+++ tests/notebooks/mirror/ipynb_to_percent/jupyter_again.py
@@ -22,4 +22,4 @@ import yaml
 print(yaml.dump(yaml.load(c)))
 
 # %%
-?next
+# ?next